A little change to css manage
