+++ src/Components/Footer/Footer.css
@@ -1,0 +1,10 @@
+.footer-content-container {
+  background-color: black;
+  color: white;
+  display: flex;
+  align-items: center;
+  justify-content: flex-end;
+  padding: 0 2rem 0 0;
+  width: 100%;
+  height: 100%;
+}
