+++ src/Components/HomeContent/HomeContent.css
@@ -1,0 +1,5 @@
+.home-content-container {
+  display: flex;
+  justify-content: center;
+  margin: 6rem 2rem 6rem 2rem;
+}
