+++ src/Components/Navbar/Navbar.css
@@ -1,0 +1,52 @@
+.navbar-content-container {
+  background-color: #dc3545;
+  color: #ffffff8c;
+  padding: 0 1rem 0 1rem;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  font-size: 17px;
+  width: 100%;
+  height: 100%;
+}
+.navbar-links-container {
+  display: flex;
+  width: 30%;
+  justify-content: space-between;
+  align-items: center;
+}
+.navbar-home {
+  font-size: 20px;
+  color: #ffc107;
+  font-weight: 600;
+}
+.navbar-options {
+  display: flex;
+  justify-content: space-around;
+  margin: 0 0 0 2rem;
+  width: 100%;
+  color: #ffffff8c;
+}
+.navbar-option-active {
+  color: white;
+}
+.navbar-search-container {
+  display: flex;
+  justify-content: space-between;
+  width: 20%;
+  height: 2.5rem;
+}
+.navbar-search-container input {
+  border-radius: 10px;
+  width: 50%;
+  margin: 0 0.5rem 0 0;
+  border: none;
+}
+.navbar-search-container button {
+  border-radius: 10px;
+  width: 50%;
+  margin: 0 0 0 0.5rem;
+  background-color: #ffc107;
+  color: black;
+  border: none;
+}
