+++ src/Components/index.js
@@ -1,3 +1,3 @@
 export { default as Footer } from "./Footer";
-export { default as NavBar } from "./Header";
+export { default as NavBar } from "./Navbar";
 export { default as HomeContent } from "./HomeContent";
