+++ src/pages/Search/index.jsx
@@ -1,0 +1,5 @@
+const Search = () => {
+  return <div>Search Page</div>;
+};
+
+export default Search;
