+++ src/pages/index.js
@@ -1,4 +1,5 @@
 export { default as Home } from "./Home";
+export { default as Search } from "./Search";
 export { default as Pokemons } from "./Pokemons";
 export { default as Favorites } from "./Favorites";
